backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
pypdf==3.17.1
google-generativeai==0.8.3
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0